chore(5): code split into logical buckets
