+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "axios": "^0.21.1",
+    "chalk": "^4.1.1",
     "date-fns": "^2.21.3",
     "form-data": "^4.0.0",
     "node-notifier": "^9.0.1",
