+++ src/data/impfstoffLookupTable.ts
@@ -1,0 +1,28 @@
+const impfstoffLookupTable = new Map([
+  [
+    "arena",
+    "https://www.doctolib.de/institut/berlin/ciz-berlin-berlin?pid=practice-158431",
+  ],
+  [
+    "tempelhof",
+    "https://www.doctolib.de/institut/berlin/ciz-berlin-berlin?pid=practice-158433",
+  ],
+  [
+    "messe",
+    "https://www.doctolib.de/institut/berlin/ciz-berlin-berlin?pid=practice-158434",
+  ],
+  [
+    "velodrom",
+    "https://www.doctolib.de/institut/berlin/ciz-berlin-berlin?pid=practice-158435",
+  ],
+  [
+    "tegel",
+    "https://www.doctolib.de/institut/berlin/ciz-berlin-berlin?pid=practice-158436",
+  ],
+  [
+    "erika",
+    "https://www.doctolib.de/institut/berlin/ciz-berlin-berlin?pid=practice-158437",
+  ],
+]);
+
+export default impfstoffLookupTable;
