+++ src/data/index.ts
@@ -1,0 +1,2 @@
+export { default as individualLinks } from "./individualLinks";
+export { default as impfstoffLookupTable } from "./impfstoffLookupTable";
