+++ src/sources/index.ts
@@ -1,0 +1,3 @@
+export { default as observePunctumMedico } from "./punctumMedico";
+export { default as observeImpfstoff } from "./impfstoff";
+export { default as observeIndividualLink } from "./individualLink";
